+++ plugin_scripts/deploy.py
@@ -4,9 +4,15 @@ import sys
 
 from gbq import BigQuery
 
-from .exceptions import DatasetSchemaDirectoryNonExistent, DeployFailed
-
 sys.tracebacklimit = 0
+
+
+class DatasetSchemaDirectoryNonExistent(Exception):
+    pass
+
+
+class DeployFailed(Exception):
+    pass
 
 
 def _validate_env_variables():
