--- plugin_scripts/exceptions.py
@@ -1,6 +1,0 @@
-class DatasetSchemaDirectoryNonExistent(Exception):
-    pass
-
-
-class DeployFailed(Exception):
-    pass
