+++ tests/test_deploy.py
@@ -1,7 +1,7 @@
 import pytest
 
 from plugin_scripts import deploy
-from plugin_scripts.exceptions import DatasetSchemaDirectoryNonExistent, DeployFailed
+from plugin_scripts.deploy import DatasetSchemaDirectoryNonExistent, DeployFailed
 
 
 @pytest.fixture
@@ -87,3 +87,28 @@ def test_main_true(mocker, gcp_project, dataset_schema_directory, credentials):
     mocker.patch("plugin_scripts.deploy.BigQuery")
     deploy.main()
     assert True
+
+
+@pytest.mark.parametrize(
+    "exception_object",
+    [
+        DatasetSchemaDirectoryNonExistent({}),
+        DeployFailed({}),
+    ],
+)
+def test_exception_init(exception_object):
+    e = exception_object
+    assert isinstance(e, Exception)
+
+
+@pytest.mark.parametrize(
+    "exception_class",
+    [
+        DatasetSchemaDirectoryNonExistent,
+        DeployFailed,
+    ],
+)
+def test_exception_throws(exception_class):
+    e = exception_class({})
+    with pytest.raises(exception_class):
+        raise e
